Move Seal to filproofs.
